refactor(news): split user allow/disallow tasks
